+++ tests/stdlib/fixtures/dump_to_sql_update_mode__update
@@ -1,4 +1,4 @@
-dump.to_sql
+dump_to_sql
 --
 {
     "tables": {
