+++ tests/stdlib/fixtures/dump_to_sql_with_updated_data
@@ -1,4 +1,4 @@
-dump.to_sql
+dump_to_sql
 --
 {
     "tables": {
